Move `signatureSpecies.ts`
